UserController and JWT added

+++ Server/Controllers/CartController.cs
@@ -1,3 +1,5 @@
+using System.Security.Claims;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Server.Data;
@@ -13,35 +15,50 @@ public class CartController(DataContext dataContext) : ControllerBase
 
     private readonly DataContext _context = dataContext;
 
+    [Authorize]
     [HttpGet]
-    public async Task<ActionResult<CartDTO>> GetCart(string customerId)
+    public async Task<ActionResult<CartDTO>> GetCart()
     {
-        var cart = await GetCartOrCreate(customerId);
+        var cart = await GetCartOrCreate();
+        if (cart == null) return Unauthorized();
         return Ok(CartToDTO(cart));
     }
 
+    [Authorize]
     [HttpPost]
-    public async Task<ActionResult<CartDTO>> AddCartItem(string customerId, int productId,int quantity){
-        var cart = await GetCartOrCreate(customerId);
+    public async Task<ActionResult<CartDTO>> AddCartItem(int productId, int quantity)
+    {
+        var cart = await GetCartOrCreate();
+        if (cart == null) return Unauthorized();
         var product = await _context.Products.FirstOrDefaultAsync(p => p.ProductId == productId);
-        if(product == null) return NotFound("The product could not be founded.");
+        if (product == null) return NotFound("The product could not be founded.");
         cart.AddCartItem(product, quantity);
         var success = await _context.SaveChangesAsync() > 0;
-        if (success) return CreatedAtAction(nameof(GetCart),CartToDTO(cart));
+        if (success) return CreatedAtAction(nameof(GetCart), CartToDTO(cart));
         return BadRequest("The item could not be added.");
     }
-
+    
+    [Authorize]
     [HttpDelete]
-    public async Task<IActionResult> DeleteCartItem(string customerId, int productId,int quantity){
-        var cart = await GetCartOrCreate(customerId);
+    public async Task<IActionResult> DeleteCartItem(int productId, int quantity)
+    {
+        var cart = await GetCartOrCreate();
+        if (cart == null) return Unauthorized();
         cart.DeleteCartItem(productId, quantity);
         var success = await _context.SaveChangesAsync() > 0;
-        if(success) return Ok();
+        if (success) return Ok(CartToDTO(cart));
         return BadRequest("The item could not be deleted.");
     }
 
-    private async Task<Cart> GetCartOrCreate(string customerId)
+    private string? GetUserId()
     {
+        return User.FindFirstValue(ClaimTypes.NameIdentifier);
+    }
+
+    private async Task<Cart?> GetCartOrCreate()
+    {
+        var customerId = GetUserId();
+        if(customerId == null) return null;
         var cart = await _context.Carts
         .Include(i => i.CartItems)
         .ThenInclude(i => i.Product)
@@ -50,19 +67,22 @@ public class CartController(DataContext dataContext) : ControllerBase
 
         if (cart == null)
         {
-            cart = new Cart {CustomerId = customerId};
+            cart = new Cart { CustomerId = customerId };
             _context.Carts.Add(cart);
             await _context.SaveChangesAsync();
         }
         return cart;
     }
 
-    private CartDTO CartToDTO(Cart cart){
-        return new CartDTO{
+    private CartDTO CartToDTO(Cart cart)
+    {
+        return new CartDTO
+        {
             CustomerId = cart.CustomerId,
             CartId = cart.CartId,
-            CartItems = cart.CartItems.Select(i => 
-                new CartItemDTO{
+            CartItems = cart.CartItems.Select(i =>
+                new CartItemDTO
+                {
                     ProductId = i.ProductId,
                     ProductImageUrl = i.Product?.ImageUrl,
                     productName = i.Product?.Name,
